Included giveaways Controller, Routes under Dashboard, ovreride registrations controller
