+++ app/controllers/dashboard/pages_controller.rb
@@ -4,5 +4,8 @@ module Dashboard
   class PagesController < Dashboard::DashboardController
     def home
     end
+
+    def tables
+    end
   end
 end
