+++ app/views/dashboard/giveaways/_footer.html.erb
@@ -13,13 +13,13 @@
 
   <% if user_signed_in? && current_user == @giveaway.user%>
     <p>
-      <%= link_to edit_dashboard_giveaway_path(@giveaway) do %>
+      <%= link_to edit_giveaway_path(@giveaway) do %>
           <i class="fas fa-edit fa-2x"></i>
       <% end %>
     </p>
 
     <p>
-      <%= link_to dashboard_giveaway_path(@giveaway), method: :delete,
+      <%= link_to giveaway_path(@giveaway), method: :delete,
         data: { confirm: t('.confirm')} do
       %>
       <i class="far fa-trash-alt fa-2x"></i>
