+++ app/views/dashboard/giveaways/_picture_fields.html.erb
