+++ app/views/dashboard/pages/home.html.erb
@@ -1,1 +1,182 @@
-Welcome
+<section class="section" id="content">
+  <%= render 'shared/flash' %>
+  <!-- Statistics -->
+  <div class="container">
+    <h3 class="title is-5">Statistics</h3>
+      <div class="columns">
+        <div class="column">
+            <div class="box statistic">
+                <div class="media ">
+                    <div class="media-left image is-48x48">
+                            <h3 class="icon title is-2 has-text-primary is-block"><i class="fas fa-folder-plus"></i></h3>
+                    </div>
+                    <div class="media-content ">
+                        <p class="subtitle is-6">Your Giveaways</p>
+                        <h4 class="title is-4 ">16</h4>
+                    </div>
+                </div>
+            </div>
+        </div>
+      <div class="column">
+          <div class="box statistic">
+              <div class="media ">
+                  <div class="media-left image is-48x48">
+                      <h3 class="icon title is-2 has-text-link is-block"><i class="fas fa-file-alt"></i></h3>
+                  </div>
+                  <div class="media-content ">
+                      <p class="subtitle is-6">Giveaways Given</p>
+                      <h4 class="title is-4 ">11</h4>
+                  </div>
+              </div>
+          </div>
+      </div>
+      <div class="column">
+          <div class="box statistic">
+              <div class="media ">
+                  <div class="media-left image is-48x48">
+                          <h3 class="icon title is-2 has-text-dark is-block"><i class="fas fa-paste"></i></h3>
+                  </div>
+                  <div class="media-content ">
+                      <p class="subtitle is-6">Your Claims</p>
+                      <h4 class="title is-4 ">6</h4>
+                  </div>
+              </div>
+          </div>
+      </div>
+      <div class="column">
+          <div class="box statistic">
+              <div class="media ">
+                  <div class="media-left image is-48x48">
+                          <h3 class="icon title is-2 has-text-danger is-block"><i class="fas fa-user-friends"></i></h3>
+                  </div>
+                  <div class="media-content ">
+                      <p class="subtitle is-6">Received</p>
+                      <h4 class="title is-4 ">2</h4>
+                  </div>
+              </div>
+          </div>
+      </div>
+  </div>
+  <!-- Panels -->
+  <h3 class="title is-5">Panels</h3>
+  <div class="columns">
+      <div class="column">
+          <div class="panel">
+              <p class="panel-heading">Create Quick Post</p>
+              <div class="panel-block">
+                  <form action="" class="is-fullwidth">
+                      <div class="field">
+                          <label for="" class="label">Title</label>
+                          <div class="control">
+                              <input type="text" name="" id="" class="input">
+                          </div>
+                      </div>
+                      <div class="field">
+                          <label for="" class="label">Category</label>
+                          <div class="select is-fullwidth">
+                              <select name="">
+                                  <option value="">Uncategorized</option>
+                              </select>
+                          </div>
+                      </div>
+                      <div class="field">
+                          <label for="" class="label">Content</label>
+                          <div class="control">
+                              <textarea type="text" name="" id="" class="textarea" rows="3"></textarea>
+                          </div>
+                      </div>
+                  </form>
+              </div>
+          </div>
+      </div>
+      <div class="column">
+          <div class="panel">
+              <p class="panel-heading">Server Status</p>
+              <div class="panel-block">
+                  <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Aut quam eos reprehenderit quos, ipsam magni. Dignissimos similique obcaecati quas repellat libero facere.</p>
+              </div>
+          </div>
+          <div class="panel">
+              <p class="panel-heading">Adminstration</p>
+              <div class="panel-block">
+                  <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Aut quam eos reprehenderit quos, ipsam magni.</p>
+              </div>
+          </div>
+      </div>
+  </div>
+  <h3 class="title is-5">Table without Panel</h3>
+  <div class="container">
+      <div class="columns">
+          <div class="column">
+              <table class="table is-striped is-bordered is-fullwidth">
+                  <thead>
+                      <tr>
+                          <th>Id</th>
+                          <th>Name</th>
+                          <th>Email</th>
+                          <th>Role</th>
+                      </tr>
+                  </thead>
+                  <tbody>
+                      <tr>
+                          <td>1</td>
+                          <td>
+                              <div class="user-media">
+                                  <img src="./images/user-image.jpg" alt="">
+                                  <div class="info">
+                                      <p class="title">Sam ZCoder</p>
+                                      <p class="subtitle">Administrator</p>
+                                  </div>
+                              </div>                         
+                          </td>
+                          <td>coder.sam@outlook.com</td>
+                          <td>Administrator</td>
+                      </tr>
+                      <tr class="is-selected">
+                          <td>2</td>
+                          <td>
+                              <div class="user-media">
+                                  <img src="./images/user-image.jpg" alt="">
+                                  <div class="info">
+                                      <p class="title">Sarah Beachoup</p>
+                                      <p class="subtitle">Moderator</p>
+                                  </div>
+                              </div>  
+                          </td>
+                          <td>sarah.beachoup@gmail.com</td>
+                          <td>Moderator</td>
+                      </tr>
+                      <tr>
+                          <td>3</td>
+                          <td>
+                              <div class="user-media">
+                                  <img src="./images/user-image.jpg" alt="">
+                                  <div class="info">
+                                      <p class="title">John Doe</p>
+                                      <p class="subtitle">Author</p>
+                                  </div>
+                              </div>
+                          </td>
+                          <td>john.doe@outlook.com</td>
+                          <td>Author</td>
+                      </tr>
+                      <tr>
+                          <td>4</td>
+                          <td>
+                              <div class="user-media">
+                                  <img src="./images/user-image.jpg" alt="">
+                                  <div class="info">
+                                      <p class="title">John Doe</p>
+                                      <p class="subtitle">User</p>
+                                  </div>
+                              </div>
+                          </td>
+                          <td>coder.sam@outlook.com</td>
+                          <td>User</td>
+                      </tr>
+                  </tbody>
+              </table>
+          </div>
+      </div>
+  </div>
+</section>
