+++ app/views/dashboard/pages/tables.html.erb
@@ -1,6 +1,3 @@
-<%= render 'partials/header'%>
-<%= render 'partials/sidebar'%>
-<%= render 'partials/page-top'%>
 <h3 class="title is-5">Tables</h3>
 <div class="columns">
     <div class="column">
@@ -145,4 +142,3 @@
         </div>
     </div>
 </div>
-<%= render 'partials/footer'%>
