+++ app/views/devise/registrations/edit.html.erb
@@ -3,7 +3,7 @@
     <article class="media">
       <div class="media-content">
         <div class="content">
-          <h2><%= t('.edit_user') %></h2>
+          <h2><%= t('.edit_account') %></h2>
         </div>
       </div>
     </article>
@@ -13,7 +13,7 @@
 
       <div class="field">
         <%= f.label t('devise.registrations.new.email') %><br />
-        <%= f.email_field :email, autofocus: true, autocomplete: "email" %>
+        <%= f.email_field :email, autofocus: true, autocomplete: "email", class:"input is-medium " %>
       </div>
 
       <% if devise_mapping.confirmable? && resource.pending_reconfirmation? %>
@@ -22,7 +22,7 @@
 
       <div class="field">
         <%= f.label t('devise.registrations.new.password')%> <i>(<%= t('.leave_blank') %>)</i><br />
-        <%= f.password_field :password, autocomplete: "new-password" %>
+        <%= f.password_field :password, autocomplete: "new-password", class:"input is-medium " %>
         <% if @minimum_password_length %>
           <br />
           <em><%= @minimum_password_length %> <%= t('.characters') %></em>
@@ -31,23 +31,23 @@
 
       <div class="field">
         <%= f.label t('.password_confirmation') %><br />
-        <%= f.password_field :password_confirmation, autocomplete: "new-password" %>
+        <%= f.password_field :password_confirmation, autocomplete: "new-password", class:"input is-medium " %>
       </div>
 
       <div class="field">
         <%= f.label t('.current_password') %> <i>(<%= t('.ppass')%>)</i><br />
-        <%= f.password_field :current_password, autocomplete: "current-password" %>
+        <%= f.password_field :current_password, autocomplete: "current-password", class:"input is-medium " %>
       </div>
 
       <div class="actions">
-        <%= f.submit t('.update') %>
+        <%= f.submit t('.update'), class:'button is-info' %>
       </div>
-    <% end %>
+    <% end %></br>
 
     <h3><%= t('.cancel') %></h3>
 
-    <p><%= t('.unhappy') %> <%= button_to t('.cancel'), registration_path(resource_name), data: { confirm: "Are you sure?" }, method: :delete %></p>
+    <p><%= t('.unhappy') %> <%= button_to t('.cancel'), registration_path(resource_name), data: { confirm: "Are you sure?" }, method: :delete, class:"button is-danger is-outlined" %></p></br>
 
-    <%= link_to t('giveaways.footer.back'), :back %>
+    <%= link_to t('giveaways.footer.back'), :back, class:"button" %>
   </div>
 </div>
