+++ app/views/giveaways/index.html.erb
@@ -1,5 +1,6 @@
 <div class="column has-text-centered">
     <% if @pagy.vars[:page].to_i <2 %>
+    
         <h1 class="subtitle is-3" style="color: #f14999;"><%= t('.welcome') %></h1>
             <div>
                 <p><%= t('.p1') %></p>
