--- app/views/layouts/dashboard/blank.html.erb
@@ -1,5 +1,0 @@
-<%= render 'partials/header'%>
-<%= render 'partials/sidebar'%>
-<%= render 'partials/page-top'%>
-<h3 class="title is-5">Blank Page</h3>
-<%= render 'partials/footer'%>
