--- app/views/layouts/dashboard/partials/_footer.html.erb
@@ -1,37 +1,0 @@
-</div>
-</section>
-<nav class="navbar is-light is-fixed-bottom" role="navigation" aria-label="Navigation" id="footer">
-    <div class="navbar-brand">
-        <a href="./index.html" class="navbar-item">
-            <img src="./images/bulma-logo.png" alt="">
-        </a>
-    </div>
-    <div class="navbar-menu">
-        <div class="navbar-start">
-            <div class="navbar-item">
-                
-                
-            </div>
-        </div>
-        <div class="navbar-end">
-            <a href="#" class="navbar-item">Dashboard</a>
-            <a href="#" class="navbar-item">Settings</a>
-            <div class="navbar-item has-dropdown is-hoverable">
-                <a href="#" class="navbar-link">Account</a>
-                <div class="navbar-dropdown">
-                    <a href="#" class="navbar-item">My Profile</a>
-                    <a href="#" class="navbar-item">Credintials</a>
-                    <a href="#" class="navbar-item">Logout</a>
-                </div>
-            </div>
-        </div>
-        
-    </div>
-</nav>
-</div>
-</section>
-
-<script defer src="https://use.fontawesome.com/releases/v5.3.1/js/all.js"></script>
-<script defer src="./javascript/main.min.js"></script>
-</body>
-</html>
