--- app/views/layouts/dashboard/partials/_header.html.erb
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Bulma Dashboard</title>
-    <link rel="stylesheet" href="./css/bulma.min.css">
-    <link rel="stylesheet" href="./css/style.min.css">
-</head>
-<body class="">
-    <section class="section is-paddingless columns" id="base_body">
