--- app/views/layouts/dashboard/partials/_hero.html.erb
@@ -1,14 +1,0 @@
-<section class="section hero is-link" id="hero">
-    <div class="media">
-        <div class="media-left">
-            <figure class="box image is-64x64">
-                <span class="icon title is-marginless is-paddingless has-text-link"><i class="fas fa-atom"></i></span>
-            </figure>
-        </div>
-        <div class="media-content">
-            <h2 class="title is-4">Welcome to Dashboard</h2>
-            <p class="subtitle is-6">Lorem ipsum dolor, sit amet consectetur adipisicing elit. Suscipit, officiis?</p>
-        </div>
-    </div>
-    <div class="is-divider"></div>
-</section>
