--- app/views/layouts/dashboard/partials/_page-top.html.erb
@@ -1,35 +1,0 @@
-<nav class="navbar is-light" role="navigation" aria-label="Navigation" id="header-nav">
-    <div class="navbar-brand">
-        <a href="./index.html" class="navbar-item">
-            <img src="./images/bulma-logo.png" alt="">
-        </a>
-    </div>
-    <div class="navbar-menu">
-        <div class="navbar-start">
-            <a href="#" class="navbar-item">Dashboard</a>
-            <a href="#" class="navbar-item">Settings</a>
-            <div class="navbar-item has-dropdown is-hoverable">
-                <a href="#" class="navbar-link">Account</a>
-                <div class="navbar-dropdown">
-                    <a href="#" class="navbar-item">My Profile</a>
-                    <a href="#" class="navbar-item">Credintials</a>
-                    <a href="#" class="navbar-item">Logout</a>
-                </div>
-            </div>
-        </div>
-        <div class="navbar-end">
-            <div class="navbar-item">
-                <div class="buttons">
-                    <a href="#" class="button is-dark">Visit Website</a>
-                    <a href="#" class="button is-danger">Logout</a>
-                </div>
-                
-            </div>
-        </div>
-    </div>
-</nav>
-
-<section class="section" id="content">
-    <div class="container">
-
-
