--- app/views/layouts/dashboard/partials/_sidebar.html.erb
@@ -1,62 +1,0 @@
-<div class="column is-one-fifth has-background-white" id="sidebar">
-    <h1 class="title is-6 is-marginless has-background-light has-text-dark">
-        
-        Bulma Dashbard
-    </h1>
-    <div class="media user">
-        <div class="media-left">
-            <figure class="image is-48x48">
-                <img class="is-rounded" src="./images/user-image.jpg" alt="">
-            </figure>
-        </div>
-        <div class="media-content">
-            <p class="title is-6">Sam ZCoder</p>
-            <p class="subtitle is-6">
-                <small>
-                    <a href="#">Profile</a> - <a href="#" class="has-text-danger">Logout</a>
-                </small>
-            </p>
-        </div>
-    </div>
-    <div class="divider"></div>
-    <aside class="menu">
-        <p class="menu-label">General</p>
-        <ul class="menu-list">
-            <li><a href="./index.html" class="is-active">
-                <span class="icon"><i class="fas fa-tachometer-alt"></i></span>
-                Dashboard
-            </a></li>
-            <li><a href="./blank.html">
-                <span class="icon"><i class="fas fa-file-alt"></i></span>
-                Blank Page
-            </a></li>
-            <li><a href="./tables.html">
-                <span class="icon"><i class="fas fa-table"></i></span>
-                Tables
-            </a></li>
-        </ul>
-        <p class="menu-label">Content</p>
-        <ul class="menu-list">
-            <li><a href="#">
-                <span class="icon"><i class="fas fa-file-alt"></i></span>
-                Pages <span class="tag is-info is-pulled-right">112</span>
-            </a></li>
-            <li><a href="#">
-                <span class="icon"><i class="fas fa-paste"></i></span>
-                Posts
-            </a></li>
-            <li class="has-dropdown">
-                <a href="#">
-                    <span class="icon"><i class="fas fa-folder-plus"></i></span>
-                    Categories
-                </a>
-                    <ul>
-                        <li><a href="#">Create new</a></li>
-                        <li><a href="#">Show All</a></li>
-                    </ul>
-
-            </li>
-        </ul>
-    </aside>
-</div>
-<div class="column is-four-fifth is-paddingless" id="page">
