+++ config/locales/en.yml
@@ -54,7 +54,7 @@ en:
     navbar:
       log_in: 'Log In'
       sign_up: 'Sign Up'
-      new_giveaway: 'Create New Giveaway'
+      new_giveaway: 'New Giveaway'
       log_out: 'Log Out'
       home: 'Home'
     footer:
@@ -106,7 +106,7 @@ en:
         pass_conf: 'Password confirmation'
         sign_up: 'Sign Up'
       edit:
-        edit_user: 'Edit User'
+        edit_account: 'Edit Account'
         leave_blank: "leave blank if you don't want to change it"
         characters: 'characters minimum'
         password_confirmation: 'Password Confirmation'
@@ -128,7 +128,7 @@ en:
 
   users:
     show:
-      edit_user: 'Edit User'
+      edit_account: 'Edit Account'
       paragraph_html: '<strong>Giveaways created by</strong>'
       paragraph2_html: '<strong>Claimed Giveaways</strong> '
